fix aspect ratio code
Remove flooding of the config-dialog list by only filling it once. The
list should never need update anyway.

Correct the logic for computing the letterbox-viewport. The old code
didn't take the backbuffer aspect into account. Now it seems to
correctly handle rendering to 4:3 windowed on 16:9 monitors.

Also add 3:2, 1.85:1, 2.39:1 and 21:9. The first three are commonly used
for film, and has just been added for fun. The latter is used in some new
HDTVs, and might actually be useful.