simplify COC-controls
The circle-of-confusion is based on physical units, but that turns
out to be difficult to control without having scenes modelled in
physical units as well. I'm not willing to start doing that, so
let's just reformulate it into focal-distance and coc-scale. The
scale factor needs to be found indevidualy for each scene through
trial and error, but whatever.