use full-screen fallback for all projection-plane intersecting spheres
Turns out, the previous code glitches in some more cases where the
occluder intersects the perspective plane. So let's use the
full-screen fallback in these cases.

While we're at it, remove that old, defined out version of the code
that is just as buggy and not in any way faster.